requirements.txt
PyQt6==6.6.1
PyOpenGL==3.1.7
numpy==1.24.3
scipy==1.11.1
pygltflib==1.16.2
PyYAML==6.0.1